+++ plugins/CMakeLists.txt
@@ -152,7 +152,7 @@ if (BUILD_SUPPORTED)
     DFHACK_PLUGIN(sort sort.cpp LINK_LIBRARIES lua)
     DFHACK_PLUGIN(steam-engine steam-engine.cpp)
     DFHACK_PLUGIN(stockflow stockflow.cpp LINK_LIBRARIES lua)
-    DFHACK_PLUGIN(stockpiles stockpiles.cpp PROTOBUFS stockpiles)
+    add_subdirectory(stockpiles)
     DFHACK_PLUGIN(stocks stocks.cpp)
     DFHACK_PLUGIN(strangemood strangemood.cpp)
     DFHACK_PLUGIN(tiletypes tiletypes.cpp Brushes.h)
