+++ plugins/stockpiles/proto/.gitignore
@@ -1,0 +1,3 @@
+*.pb.cc
+*.pb.cc.rule
+*.pb.h
