+++ plugins/stockpiles/proto/stockpiles.proto
